requirements.txt
AnilistPython
apscheduler 
PyMuPdf
aiofiles
aiohttp[speedups]
bs4
cachetools
cloudscraper
fpdf2
git+https://github.com/Ripeey/Conversation-Pyrogram
html-telegraph-poster
httpx
img2pdf
jikanpy-v4
lxml[html_clean]
m3u8
markdown2
natsort 
nhentai
opencv-python-headless
pillow
motor
pyminizip
pyrogram~=2.0.104
python-dateutil
python-dotenv
reportlab
requests
SafoneAPI
telegraph
tgcrypto
uvloop
yt-dlp